+++ src/fhir/conditions.ts
@@ -5,6 +5,7 @@ import { sortBy } from "lodash";
 import {
   flattenArrayFilters,
   searchBuilderRecords,
+  searchCommonRecords,
   searchLensRecords,
 } from "./search-helpers";
 
@@ -58,6 +59,27 @@ export async function getLensConditions(
     return filterAndSort(conditions);
   } catch (e) {
     throw new Error(`Failed fetching condition information for patient: ${e}`);
+  }
+}
+
+export async function getConditionHistory(
+  fhirClient: Client,
+  patientUPID: string
+) {
+  try {
+    const { resources: conditions } = await searchCommonRecords(
+      "Condition",
+      fhirClient,
+      {
+        patientUPID,
+        _include: ["Condition:patient", "Condition:encounter"],
+      }
+    );
+    return conditions;
+  } catch (e) {
+    throw new Error(
+      `Failed fetching condition history information for patient: ${e}`
+    );
   }
 }