+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # ctw-component-library
 
+## 0.6.0
+
+### Minor Changes
+
+- ad56045: Add ability to view condition history.
+
+## 0.5.2
+
+### Patch Changes
+
+- 98d876d: Update how style.css is imported.
+
 ## 0.5.1
 
 ### Patch Changes